BROKEN: Pushing what I've got

+++ app/models/roles.py
@@ -4,5 +4,5 @@ from app.models import db
 class Role(db.Model):
     __tablename__ = 'roles'
 
-    role_pk = db.Column(db.Integer, primary_key=True)
+    role_pk = db.Column(db.SmallInteger, primary_key=True)
     title = db.Column(db.String(32))
